removed text, found bug
it's currently set up so that, visually, if you hold the left mouse button and move your cursor onto the button, the texture will change. I want it set up so that the texture changes only if the mouse is on the button and the left mouse button has been pressed. simultaneously, in that order